% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memberstates.R
\docType{data}
\name{memberstates}
\alias{memberstates}
\title{Member states of various international organizations.}
\format{Nested list
\describe{
\item{first level}{Code for international organization. one of: coe (Council
of Europe), eu (European Union), nato (North Atlantic Treaty Organization),
oecd (Organisation for Economic Co-operation and Development)}
\item{second level}{Country code type. One of: country.name, iso3c, ios2c,
iso3n, cowc, cown}
}}
\source{
http://www.oecd.org/about/membersandpartners/list-oecd-member-countries.htm

http://www.coe.int/en/web/about-us/our-member-states

http://www.nato.int/cps/en/natolive/nato_countries.htm

https://europa.eu/european-union/about-eu/countries_en
}
\usage{
memberstates
}
\description{
Lists of the member states of various international organizations in various
country code formats. Suggest new organizations at
https://github.com/cjyetman/memberstates/issues
}
\examples{
  memberstates$nato$iso3c
  memberstates$eu$country.name

  df <- read.csv('http://data.okfn.org/data/core/country-codes/r/country-codes.csv')
  df$Capital[df$ISO3166.1.Alpha.2 \%in\% memberstates$eu$iso2c]

  url <- 'https://raw.githubusercontent.com/datasets/gdp/master/data/gdp.csv'
  df <- read.csv(url)
  df <- df[df$Year == max(df$Year), ]
  df[df$Country.Code \%in\% memberstates$eu$iso3c, ]

  library(dplyr)
  url <- 'https://raw.githubusercontent.com/datasets/gdp/master/data/gdp.csv'
  read.csv(url) \%>\%
    filter(Year == max(Year)) \%>\%
    filter(Country.Code \%in\% memberstates$eu$iso3c)
}
\keyword{datasets}

